check enabled class has been updated
