+++ Automation/src/handling_mouse_actions/ContextClick.java
@@ -1,4 +1,4 @@
-package HandlingMouseAction;
+package handling_mouse_actions;
 
 public class ContextClick {
