+++ Automation/src/handling_mouse_actions/DoubleClickOnElement.java
@@ -1,4 +1,4 @@
-package HandlingMouseAction;
+package handling_mouse_actions;
 
 import java.time.Duration;
 
@@ -8,16 +8,14 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.interactions.Actions;
 
-public class MoveToElement {
+public class DoubleClickOnElement {
 	public static void main(String[] args) {
 		WebDriver dr = new ChromeDriver();
 		dr.manage().window().maximize();
 		dr.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
 		dr.get("https://www.vtiger.com/");
-		WebElement we = dr.findElement(By.id("navbarPages"));
 		Actions a = new Actions(dr);
-		a.moveToElement(we).perform();
-		dr.quit();
+		WebElement we = dr.findElement(By.id("loginspan"));
+		a.doubleClick(we).perform();
 	}
 }
-
