+++ Automation/src/handling_mouse_actions/MoveToElement.java
@@ -1,4 +1,4 @@
-package HandlingMouseAction;
+package handling_mouse_actions;
 
 import java.time.Duration;
 
@@ -8,14 +8,16 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.interactions.Actions;
 
-public class DoubleClickOnElement {
+public class MoveToElement {
 	public static void main(String[] args) {
 		WebDriver dr = new ChromeDriver();
 		dr.manage().window().maximize();
 		dr.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
 		dr.get("https://www.vtiger.com/");
+		WebElement we = dr.findElement(By.id("navbarPages"));
 		Actions a = new Actions(dr);
-		WebElement we = dr.findElement(By.id("loginspan"));
-		a.doubleClick(we).perform();
+		a.moveToElement(we).perform();
+		dr.quit();
 	}
 }
+
